spotify: move the client behind an async trait
