+++ src/apps/mod.rs
@@ -52,7 +52,11 @@ impl Config {
             }
             spotify::app::NAME => {
                 let config = self.spotify.as_ref()?;
-                Some(Box::new(spotify::app::Spotify::new(config.clone(), input_transformer, output_transformer)))
+                Some(Box::new(spotify::app::Spotify::new(
+                    config.clone(),
+                    &spotify::client::SPOTIFY_API_CLIENT,
+                    input_transformer,
+                    output_transformer)))
             }
             youtube::app::NAME => {
                 let config = self.youtube.as_ref()?;
