+++ src/apps/spotify/client/mod.rs
@@ -1,0 +1,5 @@
+mod interface;
+pub use interface::*;
+
+mod implementation;
+pub use implementation::*;
